Move dice and edges definition files to `client-esm` folder to better reflect the core data structure
