+++ src/module/system/check/check.ts
@@ -23,7 +23,7 @@ import {
     traitSlugToObject,
 } from "@util";
 import * as R from "remeda";
-import type { Die } from "types/foundry/client/roll-term/dice-term/die.d.ts";
+import type { Die } from "types/foundry/client-esm/dice/terms/die.d.ts";
 import {
     DEGREE_OF_SUCCESS_STRINGS,
     DegreeAdjustmentsRecord,
