+++ types/foundry/client-esm/canvas/edges/collision.d.ts
