+++ types/foundry/client-esm/canvas/edges/edge.d.ts
