+++ types/foundry/client-esm/canvas/edges/edges.d.ts
