+++ types/foundry/client-esm/canvas/edges/module.d.ts
