+++ types/foundry/client-esm/canvas/edges/vertex.d.ts
