+++ types/foundry/client-esm/dice/module.d.ts
@@ -1,0 +1,1 @@
+export * as terms from "./terms/module.ts";
