+++ types/foundry/client-esm/dice/terms/die.d.ts
@@ -1,4 +1,4 @@
-import type { DiceTerm } from "./base.d.ts";
+import type { DiceTerm } from "./dice-term.d.ts";
 
 /**
  * Define a fair n-sided die term that can be used as part of a Roll formula
