+++ types/foundry/client-esm/dice/terms/module.d.ts
@@ -1,0 +1,12 @@
+export * from "./coin.ts";
+export * from "./dice-term.ts";
+export * from "./die.ts";
+export * from "./fate-die.ts";
+
+export * from "./function-term.ts";
+export * from "./numeric-term.ts";
+export * from "./operator-term.ts";
+export * from "./parenthetical-term.ts";
+export * from "./pool-term.ts";
+export * from "./roll-term.ts";
+export * from "./string-term.ts";
