+++ types/foundry/client-esm/dice/terms/roll-term.d.ts
