+++ types/foundry/client/pixi/module.d.ts
@@ -1,3 +1,3 @@
-export * as edges from "./edges/module.ts";
+export * as edges from "../../client-esm/canvas/edges/module.ts";
 export * as sources from "../../client-esm/canvas/sources/module.ts";
 export * from "./scene-manager.ts";
