--- types/foundry/client/roll-term/index.d.ts
@@ -1,11 +1,0 @@
-import "./base.d.ts";
-import "./dice-term/base.d.ts";
-import "./dice-term/coin.d.ts";
-import "./dice-term/die.d.ts";
-import "./dice-term/fate-die.d.ts";
-import "./function-term.js";
-import "./numeric-term.d.ts";
-import "./operator-term.d.ts";
-import "./parenthetical-term.d.ts";
-import "./pool-term.d.ts";
-import "./string-term.d.ts";
