--- types/foundry/client/roll-term/module.d.ts
@@ -1,1 +1,0 @@
-export * as terms from "./terms.ts";
