--- types/foundry/client/roll-term/terms.d.ts
@@ -1,12 +1,0 @@
-export * from "./dice-term/base.js";
-export * from "./dice-term/coin.js";
-export * from "./dice-term/die.js";
-export * from "./dice-term/fate-die.js";
-
-export * from "./base.js";
-export * from "./function-term.js";
-export * from "./numeric-term.js";
-export * from "./operator-term.js";
-export * from "./parenthetical-term.js";
-export * from "./pool-term.js";
-export * from "./string-term.js";
